+++ rust/blockstore/src/arrow/block/key/f32_key.rs
@@ -1,10 +1,12 @@
-use super::delta::BlockKeyArrowBuilder;
-use crate::arrow::types::{ArrowReadableKey, ArrowReadableValue, ArrowWriteableKey};
-use arrow::array::{Array, Float32Builder, StringBuilder, UInt32Array, UInt32Builder};
+use crate::arrow::{
+    block::delta::{BlockDelta, BlockKeyArrowBuilder},
+    types::{ArrowReadableKey, ArrowReadableValue, ArrowWriteableKey},
+};
+use arrow::array::{Array, Float32Array, Float32Builder, StringBuilder};
 use std::sync::Arc;
 
-impl ArrowWriteableKey for u32 {
-    type ReadableKey<'referred_data> = u32;
+impl ArrowWriteableKey for f32 {
+    type ReadableKey<'referred_data> = f32;
 
     fn offset_size(_: usize) -> usize {
         0
@@ -15,16 +17,16 @@ impl ArrowWriteableKey for u32 {
         _: usize,
     ) -> BlockKeyArrowBuilder {
         let prefix_builder = StringBuilder::with_capacity(item_count, prefix_capacity);
-        let key_builder = UInt32Builder::with_capacity(item_count);
-        BlockKeyArrowBuilder::UInt32((prefix_builder, key_builder))
+        let key_builder = Float32Builder::with_capacity(item_count);
+        BlockKeyArrowBuilder::Float32((prefix_builder, key_builder))
     }
 }
 
-impl ArrowReadableKey<'_> for u32 {
+impl ArrowReadableKey<'_> for f32 {
     fn get(array: &Arc<dyn Array>, index: usize) -> Self {
         array
             .as_any()
-            .downcast_ref::<UInt32Array>()
+            .downcast_ref::<Float32Array>()
             .unwrap()
             .value(index)
     }
@@ -33,7 +35,7 @@ impl ArrowReadableKey<'_> for u32 {
         prefix: &str,
         key: Self,
         value: V,
-        delta: &mut super::delta::BlockDelta,
+        delta: &mut BlockDelta,
     ) {
         V::add_to_delta(prefix, key, value, delta);
     }
