+++ rust/blockstore/src/arrow/block/mod.rs
@@ -1,13 +1,10 @@
-mod bool_key;
 mod data_record_value;
 pub(in crate::arrow) mod delta;
-mod f32_key;
 mod int32array_value;
+mod key;
 mod roaring_bitmap_value;
-mod str_key;
 mod str_value;
 mod types;
-mod u32_key;
 mod u32_value;
 // Re-export types at the arrow_blockfile module level
 pub(crate) use types::*;
